// front/src/app/api-client/model/routine-history-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Routines terminées d\'un jour (historique)
 */
export interface RoutineHistoryDTO { 
    date?: string;
    /**
     * Ids des routines terminées ce jour-là
     */
    routinesTerminees?: Array<number>;
}